requirements.txt
#rioxarray
progressbar

# File Downloader
requests
tqdm
bs4
numpy
tifffile
pyyaml
matplotlib
rioxarray